+++ packages/runtime/tests/runtime.rs
@@ -3,17 +3,14 @@ use std::{collections::HashMap, sync::Once};
 use lagon_runtime::{
     http::{Method, Request, Response, RunResult},
     isolate::{Isolate, IsolateOptions},
+    runtime::{Runtime, RuntimeOptions},
 };
 
 fn setup() {
     static START: Once = Once::new();
 
     START.call_once(|| {
-        let platform = v8::new_default_platform(0, false).make_shared();
-        v8::V8::initialize_platform(platform);
-        v8::V8::initialize();
-
-        v8::V8::set_flags_from_string("--disallow-code-generation-from-strings");
+        Runtime::new(RuntimeOptions::default());
     });
 }
 
@@ -28,12 +25,14 @@ async fn execute_function() {
     ));
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "Hello world".into(),
             headers: None,
@@ -60,12 +59,14 @@ async fn environment_variables() {
     );
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "Hello world".into(),
             headers: None,
@@ -85,12 +86,14 @@ async fn get_body() {
     ));
 
     assert_eq!(
-        isolate.run(Request {
-            body: "Hello world".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "Hello world".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "Hello world".into(),
             headers: None,
@@ -110,12 +113,14 @@ async fn get_input() {
     ));
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "https://hello.world".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "https://hello.world".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "https://hello.world".into(),
             headers: None,
@@ -135,12 +140,14 @@ async fn get_method() {
     ));
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::POST,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::POST,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "POST".into(),
             headers: None,
@@ -163,12 +170,14 @@ async fn get_headers() {
     headers.insert("x-auth".into(), "token".into());
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers,
-            method: Method::POST,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers,
+                method: Method::POST,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "token".into(),
             headers: None,
@@ -197,12 +206,14 @@ async fn return_headers() {
     headers.insert("X-Test".into(), "test".into());
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "Hello world".into(),
             headers: Some(headers),
@@ -231,12 +242,14 @@ async fn return_headers_from_headers_api() {
     headers.insert("X-Test".into(), "test".into());
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "Hello world".into(),
             headers: Some(headers),
@@ -258,16 +271,47 @@ async fn return_status() {
     ));
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Response(Response {
             body: "Moved permanently".into(),
             headers: None,
             status: 302,
+        })
+    );
+}
+
+#[tokio::test]
+async fn return_uint8array() {
+    setup();
+    let mut isolate = Isolate::new(IsolateOptions::new(
+        "export function handler() {
+    // TextEncoder#encode returns a Uint8Array
+    const body = new TextEncoder().encode('Hello world');
+    return new Response(body);
+}"
+        .into(),
+    ));
+
+    assert_eq!(
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
+        RunResult::Response(Response {
+            body: "Hello world".into(),
+            headers: None,
+            status: 200,
         })
     );
 }
@@ -284,12 +328,14 @@ async fn timeout_reached() {
     ));
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::Timeout(),
     );
 }
@@ -317,12 +363,14 @@ async fn memory_reached() {
     ); // Increase timeout for CI
 
     assert_eq!(
-        isolate.run(Request {
-            body: "".into(),
-            headers: HashMap::new(),
-            method: Method::GET,
-            url: "".into(),
-        }),
+        isolate
+            .run(Request {
+                body: "".into(),
+                headers: HashMap::new(),
+                method: Method::GET,
+                url: "".into(),
+            })
+            .0,
         RunResult::MemoryLimit(),
     );
 }