+++ packages/serverless/src/main.rs
@@ -4,13 +4,14 @@ use hyper::{Body, Request as HyperRequest, Response as HyperResponse, Server};
 use lagon_runtime::http::RunResult;
 use lagon_runtime::isolate::{Isolate, IsolateOptions};
 use lagon_runtime::runtime::{Runtime, RuntimeOptions};
-use mysql::Pool;
+use metrics::{counter, histogram, increment_counter};
+use metrics_exporter_prometheus::PrometheusBuilder;
+use mysql::{Opts, OptsBuilder, Pool, SslOpts};
 use s3::creds::Credentials;
 use s3::Bucket;
 use std::collections::HashMap;
 use std::convert::Infallible;
 use std::net::SocketAddr;
-use std::time::Instant;
 use tokio::task::LocalSet;
 
 use crate::deployments::assets::handle_asset;
@@ -18,17 +19,17 @@ use crate::deployments::filesystem::get_deployment_code;
 use crate::deployments::get_deployments;
 use crate::deployments::pubsub::listen_pub_sub;
 use crate::http::response_to_hyper_response;
+use crate::logger::init_logger;
 
 mod deployments;
 mod http;
+mod logger;
 
 async fn handle_request(
     req: HyperRequest<Body>,
     request_tx: flume::Sender<HyperRequest<Body>>,
     response_rx: flume::Receiver<RunResult>,
 ) -> Result<HyperResponse<Body>, Infallible> {
-    let now = Instant::now();
-
     request_tx.send_async(req).await;
 
     let result = response_rx
@@ -38,13 +39,6 @@ async fn handle_request(
 
     match result {
         RunResult::Response(response) => {
-            // println!(
-            //     "Response: {:?} in {:?} (CPU time) (Total: {:?})",
-            //     response,
-            //     duration,
-            //     now.elapsed()
-            // );
-
             let response = response_to_hyper_response(response);
 
             Ok(response)
@@ -64,9 +58,11 @@ async fn handle_request(
 
 #[tokio::main]
 async fn main() {
-    dotenv::dotenv().ok();
+    dotenv::dotenv().expect("Failed to load .env file");
+    init_logger().expect("Failed to init logger");
+
     let runtime = Runtime::new(RuntimeOptions::default());
-    let addr = SocketAddr::from(([127, 0, 0, 1], 4000));
+    let addr = SocketAddr::from(([0, 0, 0, 0], 4000));
 
     let (request_tx, request_rx) = flume::unbounded::<HyperRequest<Body>>();
     let (response_tx, response_rx) = flume::unbounded::<RunResult>();
@@ -82,9 +78,17 @@ async fn main() {
         }
     }));
 
+    let builder = PrometheusBuilder::new();
+    builder.install().expect("Failed to start metrics exporter");
+
     let url = dotenv::var("DATABASE_URL").expect("DATABASE_URL must be set");
     let url = url.as_str();
-    let pool = Pool::new(url).unwrap();
+    let opts = Opts::from_url(url).expect("Failed to parse DATABASE_URL");
+    #[cfg(not(debug_assertions))]
+    let opts = OptsBuilder::from_opts(opts).ssl_opts(Some(
+        SslOpts::default().with_danger_accept_invalid_certs(true),
+    ));
+    let pool = Pool::new(opts).unwrap();
     let conn = pool.get_conn().unwrap();
 
     let bucket_name = dotenv::var("S3_BUCKET").expect("S3_BUCKET must be set");
@@ -111,18 +115,28 @@ async fn main() {
             let mut isolates = HashMap::new();
 
             loop {
-                let request = request_rx.recv_async().await.unwrap();
-                let request = hyper_request_to_request(request).await;
+                let hyper_request = request_rx.recv_async().await.unwrap();
+
+                let mut url = hyper_request.uri().to_string();
+                // Remove the leading '/' from the url
+                url.remove(0);
+
+                let request = hyper_request_to_request(hyper_request).await;
 
                 let hostname = request.headers.get("host").unwrap().clone();
                 let deployments = deployments.read().await;
 
                 match deployments.get(&hostname) {
                     Some(deployment) => {
-                        let url = &mut request.url.clone();
-                        url.remove(0);
+                        let labels = vec![
+                            ("deployment", deployment.id.clone()),
+                            ("function", deployment.function_id.clone()),
+                        ];
 
-                        if let Some(asset) = deployment.assets.iter().find(|asset| *asset == url) {
+                        increment_counter!("lagon_requests", &labels);
+                        counter!("lagon_bytes_in", request.len() as u64, &labels);
+
+                        if let Some(asset) = deployment.assets.iter().find(|asset| *asset == &url) {
                             // TODO: handle read error
                             let response = handle_asset(deployment, asset).unwrap();
                             let response = RunResult::Response(response);
@@ -142,9 +156,22 @@ async fn main() {
                                 Isolate::new(options)
                             });
 
-                            let result = isolate.run(request);
+                            let (run_result, maybe_statistics) = isolate.run(request);
 
-                            response_tx.send_async(result).await.unwrap();
+                            if let Some(statistics) = maybe_statistics {
+                                histogram!("lagon_isolate_cpu_time", statistics.cpu_time, &labels);
+                                histogram!(
+                                    "lagon_isolate_memory_usage",
+                                    statistics.memory_usage as f64,
+                                    &labels
+                                );
+                            }
+
+                            if let RunResult::Response(response) = &run_result {
+                                counter!("lagon_bytes_out", response.len() as u64, &labels);
+                            }
+
+                            response_tx.send_async(run_result).await.unwrap();
                         }
                     }
                     None => {