+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lagon/runtime
 
+## 0.1.2
+
+### Patch Changes
+
+- [#156](https://github.com/lagonapp/lagon/pull/156) [`dcfdf5d`](https://github.com/lagonapp/lagon/commit/dcfdf5d591fb787a8d9c549345f8c8901593a455) Thanks [@QuiiBz](https://github.com/QuiiBz)! - Add cpu and memory statistics
+
 ## 0.1.1
 
 ### Patch Changes